requirements.txt
numpy
pygame
opencv-python
pillow
